cloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
nltk
python_dateutil
dateparser
flask
